fix makefiles and docs for go mod update and flogo v0.9.4

+++ hyperledger-fabric/samples/abac/start-fn.sh
@@ -1,13 +1,16 @@
 #!/bin/bash
 # restart fabric sample first-network, including CA servers without TLS
 
-FABRIC_SAMPLE_PATH=${GOPATH}/src/github.com/hyperledger/fabric-samples/first-network
-cd ${FABRIC_SAMPLE_PATH}
+FAB_PATH=${1:-"${GOPATH}/src/github.com/hyperledger/fabric-samples"}
+cd ${FAB_PATH}/first-network
 
-# cleanup started network if any
-./byfn.sh down
-docker rm $(docker ps -a | grep dev-peer | awk '{print $1}')
-docker rmi $(docker images | grep dev-peer | awk '{print $3}')
+cons=$(docker ps -a -f name=example.com | wc -l)
+if [ $cons -gt 1 ]; then
+  echo "cleanup previous running containers ..."
+  ./byfn.sh down
+  docker rm $(docker ps -a | grep dev-peer | awk '{print $1}')
+  docker rmi $(docker images | grep dev-peer | awk '{print $3}')
+fi
 
 # turn off TLS on CA servers
 cp docker-compose-ca.yaml docker-compose-ca.yaml.orig
